put chain unit tests in a separate file
